corrected according to comments

+++ lab1/main.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
-#include "include/number1.h"
+#include "number1.h"
 
 int main()
 {
     int start, end;
     std::cin >> start >> end;
-    int ans = how_one_in_cons(start, end);
+    uint64_t ans = how_one_in_cons(start, end);
     std::cout << ans << std::endl;
     return 0;
 }
